Do not throw Exception when task is skipped
Currently, when we define our own 'onlyIf' spec and run gradle in offline mode and the file to download does not exist,
an exception is thrown even if the 'onlyIf' check defined by the user would return false.
By moving the whole check into the actual task, it does not fail prematurely.